add integration tests using GPU for osc calculations
