+++ tests/test-DP-propagator.cpp
@@ -51,35 +51,35 @@ TEST_P(DPpropagatorTest /*unused*/, CompareToNuFast_sinSquaredThetas /*unused*/)
 TEST_P(DPpropagatorTest /*unused*/, autogradTest /*unused*/)
 {
 
-    _setParamValues(/*forceLowerOctant=*/false, /*interpretSinSquaredThetas=*/false);
-
-    theta23tensor.requiresGrad(true);
-
-    Tensor pmnsTensor = pmns.build();
-    tensorPropagator.setMixingMatrix(pmnsTensor);
-
-    // get Propagator probabilities
-    Tensor probabilities = tensorPropagator.calculateProbs();
-    Tensor muSurvivalProb = probabilities.getValues({0, 1, 1});
-
-    muSurvivalProb.backward();
-
-    NT_INFO("Propagator:   d P_(mu->mu) / d theta_23 = {}", pmns.getTheta23Tensor().grad().getValue<float>());
-
-    // get DPpropagator probabilities
-    Tensor dpProbabilities = dpPropagator.calculateProbs();
-    Tensor dpMuSurvivalProb = dpProbabilities.getValues({0, 1, 1});
-
-    dpMuSurvivalProb.backward();
-
-    NT_INFO("DPpropagator: d P_(mu->mu) / d theta_23 = {}", theta23tensor.grad().getValue<float>());
-
-    // check that the values are close to each other
-    ASSERT_NEAR(pmns.getTheta23Tensor().grad().getValue<float>(), theta23tensor.grad().getValue<float>(), tolerance);
+    autogradTest();
 }
 
-INSTANTIATE_TEST_CASE_P(OscProb, DPpropagatorTest,
-                        ::testing::Values(-M_PI, -0.8 * M_PI, -0.5 * M_PI, -0.2 * M_PI, 0.0, 0.3 * M_PI, 0.5 * M_PI,
-                                          0.7 * M_PI, M_PI));
+#if COMPILE_GPU_TESTS
 
+INSTANTIATE_TEST_CASE_P(
+    OscProb, DPpropagatorTest,
+    ::testing::Values(std::make_tuple(-M_PI, dtypes::kCPU), std::make_tuple(-0.8 * M_PI, dtypes::kCPU),
+                      std::make_tuple(-0.6 * M_PI, dtypes::kCPU), std::make_tuple(-0.4 * M_PI, dtypes::kCPU),
+                      std::make_tuple(-0.2 * M_PI, dtypes::kCPU), std::make_tuple(0.0, dtypes::kCPU),
+                      std::make_tuple(0.2 * M_PI, dtypes::kCPU), std::make_tuple(0.4 * M_PI, dtypes::kCPU),
+                      std::make_tuple(0.6 * M_PI, dtypes::kCPU), std::make_tuple(0.8 * M_PI, dtypes::kCPU),
+                      std::make_tuple(M_PI, dtypes::kCPU), std::make_tuple(-M_PI, dtypes::kCPU),
+                      std::make_tuple(-0.8 * M_PI, dtypes::kGPU), std::make_tuple(-0.6 * M_PI, dtypes::kGPU),
+                      std::make_tuple(-0.4 * M_PI, dtypes::kGPU), std::make_tuple(-0.2 * M_PI, dtypes::kGPU),
+                      std::make_tuple(0.0, dtypes::kGPU), std::make_tuple(0.2 * M_PI, dtypes::kGPU),
+                      std::make_tuple(0.4 * M_PI, dtypes::kGPU), std::make_tuple(0.6 * M_PI, dtypes::kGPU),
+                      std::make_tuple(0.8 * M_PI, dtypes::kGPU), std::make_tuple(M_PI, dtypes::kGPU)));
+
+#else
+
+INSTANTIATE_TEST_CASE_P(
+    OscProb, DPpropagatorTest,
+    ::testing::Values(std::make_tuple(-M_PI, dtypes::kCPU), std::make_tuple(-0.8 * M_PI, dtypes::kCPU),
+                      std::make_tuple(-0.6 * M_PI, dtypes::kCPU), std::make_tuple(-0.4 * M_PI, dtypes::kCPU),
+                      std::make_tuple(-0.2 * M_PI, dtypes::kCPU), std::make_tuple(0.0, dtypes::kCPU),
+                      std::make_tuple(0.2 * M_PI, dtypes::kCPU), std::make_tuple(0.4 * M_PI, dtypes::kCPU),
+                      std::make_tuple(0.6 * M_PI, dtypes::kCPU), std::make_tuple(0.8 * M_PI, dtypes::kCPU),
+                      std::make_tuple(M_PI, dtypes::kCPU)));
+
+#endif
 // NOLINTEND(readability-magic-numbers, cppcoreguidelines-avoid-magic-numbers)
